update wakepy._system
- Move system related constant from .constants to ._system
- Remove everything else.
